+++ lib/app/models/enums.dart
@@ -249,7 +249,6 @@ extension FullPlayerActionsLabel on FullPlayerActions {
   }
 }
 
-// TODO: add cast
 // TODO: add bookmarks
 
 const defaultPlayerActions = <FullPlayerActions>[