+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Customizable full player action buttons via "more"
 
+### Fixed
+
+- Player stop with mini player dismiss race condition
+
 ## [v0.5.5] - 2026-07-09
 
 ### Changed